Reestruture o projeto
