+++ MJVSchool.NET-CRUDdaJustica.sln
@@ -11,7 +11,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CrudDaJustica.Cli.App", "sr
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CrudDaJustica.Cli.Lib", "src\CrudDaJustica.Cli.Lib\CrudDaJustica.Cli.Lib.csproj", "{CEB49CAB-EE7B-4A90-9B2F-1781C4407E52}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CrudDaJustica.Cli.App.Test", "test\CrudDaJustica.Cli.App.Tests\CrudDaJustica.Cli.App.Test.csproj", "{220BD4BB-6BFD-461D-A762-46E6AEEB021C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CrudDaJustica.Data.Lib", "src\CrudDaJustica.Data.Lib\CrudDaJustica.Data.Lib.csproj", "{5963C2EA-1C32-4091-8D57-A51C351781D3}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CrudDaJustica.Data.Lib.Test", "test\CrudDaJustica.Data.Lib.Tests\CrudDaJustica.Data.Lib.Test.csproj", "{1EA33C49-3014-408B-B743-6B1C6FC822B8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,10 +29,14 @@ Global
 		{CEB49CAB-EE7B-4A90-9B2F-1781C4407E52}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CEB49CAB-EE7B-4A90-9B2F-1781C4407E52}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CEB49CAB-EE7B-4A90-9B2F-1781C4407E52}.Release|Any CPU.Build.0 = Release|Any CPU
-		{220BD4BB-6BFD-461D-A762-46E6AEEB021C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{220BD4BB-6BFD-461D-A762-46E6AEEB021C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{220BD4BB-6BFD-461D-A762-46E6AEEB021C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{220BD4BB-6BFD-461D-A762-46E6AEEB021C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5963C2EA-1C32-4091-8D57-A51C351781D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5963C2EA-1C32-4091-8D57-A51C351781D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5963C2EA-1C32-4091-8D57-A51C351781D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5963C2EA-1C32-4091-8D57-A51C351781D3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1EA33C49-3014-408B-B743-6B1C6FC822B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1EA33C49-3014-408B-B743-6B1C6FC822B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1EA33C49-3014-408B-B743-6B1C6FC822B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1EA33C49-3014-408B-B743-6B1C6FC822B8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -38,7 +44,8 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{163B1806-63E9-4893-A7DA-12F0D9E75C07} = {49CE7C7B-6A2F-4AA6-917C-ABAE226F3723}
 		{CEB49CAB-EE7B-4A90-9B2F-1781C4407E52} = {49CE7C7B-6A2F-4AA6-917C-ABAE226F3723}
-		{220BD4BB-6BFD-461D-A762-46E6AEEB021C} = {476009DC-30F7-49CC-93A8-14F7F1B040CA}
+		{5963C2EA-1C32-4091-8D57-A51C351781D3} = {49CE7C7B-6A2F-4AA6-917C-ABAE226F3723}
+		{1EA33C49-3014-408B-B743-6B1C6FC822B8} = {476009DC-30F7-49CC-93A8-14F7F1B040CA}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {CDCA97D1-0179-436E-AE45-2EF603387F58}
