+++ src/CrudDaJustica.Cli.App/Controller/HeroController.cs
@@ -1,7 +1,7 @@
-﻿using CrudDaJustica.Cli.App.Data;
-using CrudDaJustica.Cli.App.Model;
-using CrudDaJustica.Cli.App.Services;
-using CrudDaJustica.Cli.App.View.Models;
+﻿using CrudDaJustica.Cli.App.View.Models;
+using CrudDaJustica.Data.Lib.Data;
+using CrudDaJustica.Data.Lib.Model;
+using CrudDaJustica.Data.Lib.Services;
 using System.Globalization;
 
 namespace CrudDaJustica.Cli.App.Controller;
